+++ zetaclient/chains/ton/observer/inbound.go
@@ -15,7 +15,7 @@ import (
 	"github.com/zeta-chain/node/x/crosschain/types"
 	"github.com/zeta-chain/node/zetaclient/chains/ton/encoder"
 	"github.com/zeta-chain/node/zetaclient/compliance"
-	zetaclientconfig "github.com/zeta-chain/node/zetaclient/config"
+	"github.com/zeta-chain/node/zetaclient/config"
 	"github.com/zeta-chain/node/zetaclient/logs"
 	"github.com/zeta-chain/node/zetaclient/zetacore"
 )
@@ -170,11 +170,36 @@ func (ob *Observer) ProcessInboundTrackers(ctx context.Context) error {
 		return err
 	}
 
+	return ob.observeInboundTrackers(ctx, trackers, false)
+}
+
+// ProcessInternalTrackers processes internal inbound trackers
+func (ob *Observer) ProcessInternalTrackers(ctx context.Context) error {
+	trackers := ob.GetInboundInternalTrackers(ctx)
+	if len(trackers) > 0 {
+		ob.Logger().Inbound.Info().Int("total_count", len(trackers)).Msg("processing internal trackers")
+	}
+
+	return ob.observeInboundTrackers(ctx, trackers, true)
+}
+
+// observeInboundTrackers observes given inbound trackers
+func (ob *Observer) observeInboundTrackers(
+	ctx context.Context,
+	trackers []types.InboundTracker,
+	isInternal bool,
+) error {
+	// take at most MaxInternalTrackersPerScan for each scan
+	if len(trackers) > config.MaxInboundTrackersPerScan {
+		trackers = trackers[:config.MaxInboundTrackersPerScan]
+	}
+
 	logSkippedTracker := func(hash string, reason string, err error) {
 		ob.Logger().Inbound.Warn().
 			Err(err).
 			Str(logs.FieldTx, hash).
 			Str("reason", reason).
+			Bool("is_internal", isInternal).
 			Msg("skipping inbound tracker")
 	}
 
@@ -282,7 +307,8 @@ func (ob *Observer) voteInbound(ctx context.Context, tx *toncontracts.Transactio
 
 	logger = ob.Logger().Inbound
 	msg := inbound.intoVoteMessage(operatorAddress, senderChain, zetaChain)
-	_, err = ob.ZetaRepo().VoteInbound(ctx, logger, msg, zetacore.PostVoteInboundExecutionGasLimit)
+	_, err = ob.ZetaRepo().
+		VoteInbound(ctx, logger, msg, zetacore.PostVoteInboundExecutionGasLimit, ob.WatchMonitoringError)
 	if err != nil {
 		return err
 	}
@@ -364,7 +390,7 @@ func newInbound(tx *toncontracts.Transaction) (*Inbound, error) {
 }
 
 func (inbound *Inbound) isCompliant() bool {
-	return !zetaclientconfig.ContainRestrictedAddress(
+	return !config.ContainRestrictedAddress(
 		inbound.receiver.Hex(),
 		inbound.sender.ToRaw(),
 		inbound.sender.ToHuman(false, false),