+++ changelog.md
@@ -53,6 +53,7 @@ Also EnableSolanaAddressLookupTable feature flag should be set.
 
 * [4153](https://github.com/zeta-chain/node/pull/4153) - make the gas limit used for gateway calls a configurable parameter
 * [4277](https://github.com/zeta-chain/node/pull/4277) - add dry-mode support for TON
+* [4295](https://github.com/zeta-chain/node/pull/4295) - monitor failed inbound votes and add them to an internal inbound tracker cache
 
 ### Refactor